+++ src/hello_imgui/internal/backend_impls/backend_window_helper/glfw_window_helper.cpp
@@ -3,6 +3,11 @@
 #include "window_geometry_helper.h"
 #include "GLFW/glfw3.h"
 
+#ifdef __EMSCRIPTEN__
+#include <GLFW/emscripten_glfw3.h>
+#endif
+
+#include "hello_imgui/internal/platform/platform_detection.h"
 
 namespace HelloImGui { namespace BackendApi
 {
@@ -133,22 +138,44 @@ namespace HelloImGui { namespace BackendApi
             glfwWindowHint(GLFW_CLIENT_API, GLFW_NO_API);
         }
 
+#ifndef __EMSCRIPTEN__
         if (appWindowParams.borderless)
             glfwWindowHint(GLFW_DECORATED, GLFW_FALSE);
         else
             glfwWindowHint(GLFW_DECORATED, GLFW_TRUE);
+#endif
 
         if (appWindowParams.resizable)
             glfwWindowHint(GLFW_RESIZABLE, GLFW_TRUE);
         else
             glfwWindowHint(GLFW_RESIZABLE, GLFW_FALSE);
 
+#ifdef HELLOIMGUI_IS_DESKTOP_PLATFORM
+        if (appWindowParams.topMost)
+            glfwWindowHint(GLFW_FLOATING, GLFW_TRUE);
+        else
+            glfwWindowHint(GLFW_FLOATING, GLFW_FALSE);
+#endif
+
         glfwWindowHint(GLFW_VISIBLE, GLFW_FALSE);
 
         // GLFW_SCALE_TO_MONITOR is set to false: we set manually the window size in screen coordinates,
         // then AbstractRunner::MakeWindowSizeRelativeTo96Ppi_IfRequired() may resize the window at the second frame
         // (depending on the monitor on which it is placed)
+#ifndef __EMSCRIPTEN__
+        // GLFW_SCALE_TO_MONITOR is deprecated for emscripten
+        // (and if set to true, the rendering is too big)
         glfwWindowHint(GLFW_SCALE_TO_MONITOR, GLFW_FALSE);
+#endif
+
+#ifdef __EMSCRIPTEN__
+        // Make hi dpi aware (note that GLFW_SCALE_FRAMEBUFFER already defaults to GLFW_TRUE
+        // so technically this is not necessary)
+        glfwWindowHint(GLFW_SCALE_FRAMEBUFFER, GLFW_TRUE);
+
+        // setting the association window <-> canvas
+        emscripten_glfw_set_next_window_canvas_selector("#canvas");
+#endif
 
         auto window = glfwCreateWindow(
             windowSize[0], windowSize[1],
@@ -158,6 +185,11 @@ namespace HelloImGui { namespace BackendApi
         );
         if (window == nullptr)
             BACKEND_THROW("BackendGlfw::CreateWindow / glfwCreateWindow failed");
+
+#ifdef __EMSCRIPTEN__
+        // makes the canvas resizable and match the full window size
+        emscripten_glfw_make_canvas_resizable(window, "window", nullptr);
+#endif
 
         if (appWindowParams.windowGeometry.windowSizeState == WindowSizeState::Minimized)
             glfwIconifyWindow(window);
@@ -235,8 +267,10 @@ namespace HelloImGui { namespace BackendApi
     {
         auto glfwWindow = (GLFWwindow *)(window);
         glfwShowWindow(glfwWindow);
+#ifdef HELLOIMGUI_IS_DESKTOP_PLATFORM
         glfwFocusWindow(glfwWindow);
         glfwRequestWindowAttention(glfwWindow);
+#endif
     }
 
     ScreenBounds GlfwWindowHelper::GetWindowBounds(WindowPointer window)
@@ -269,7 +303,7 @@ namespace HelloImGui { namespace BackendApi
 
     float GlfwWindowHelper::GetWindowSizeDpiScaleFactor(WindowPointer window)
     {
-#ifdef __APPLE__
+#if defined(__APPLE__) || defined(__EMSCRIPTEN__)
         return 1.f;
 #else
         ImVec2 scale = _GetWindowContentScale(window);
@@ -289,6 +323,13 @@ namespace HelloImGui { namespace BackendApi
     {
         int visible = glfwGetWindowAttrib((GLFWwindow *) window, GLFW_VISIBLE);
         return (visible == 0);
+    }
+
+    void GlfwWindowHelper::SetWindowTopMost(WindowPointer window, bool topMost)
+    {
+#ifdef HELLOIMGUI_IS_DESKTOP_PLATFORM
+        glfwSetWindowAttrib((GLFWwindow *) window, GLFW_FLOATING, topMost ? GLFW_TRUE : GLFW_FALSE);
+#endif
     }
 
 //    void truc(WindowPointer window)