+++ src/hello_imgui/internal/backend_impls/backend_window_helper/sdl_window_helper.cpp
@@ -12,6 +12,8 @@
 #include "hello_imgui/internal/platform/getAppleBundleResourcePath.h"
 #endif
 
+#include "hello_imgui/internal/platform/platform_detection.h"
+
 
 namespace HelloImGui { namespace BackendApi
 {
@@ -161,6 +163,12 @@ namespace HelloImGui { namespace BackendApi
         SDL_GetWindowPosition(window, &windowPosition[0], &windowPosition[1]);
         SDL_GetWindowSize(window, &windowSize[0], &windowSize[1]);
 
+        // Set topMost if requested (only on desktop platforms)
+#ifdef HELLOIMGUI_IS_DESKTOP_PLATFORM
+        if (appWindowParams.topMost)
+            SDL_SetWindowAlwaysOnTop(window, SDL_TRUE);
+#endif
+
         // printf("Final window size: %ix%i\n", windowSize[0], windowSize[1]);
         // printf("Final window position: %ix%i\n", windowPosition[0], windowPosition[1]);
 
@@ -196,8 +204,10 @@ namespace HelloImGui { namespace BackendApi
         // Despite those efforts, the app does not come to the front under MacOS
         auto sdlWindow = (SDL_Window *)(window);
         SDL_ShowWindow(sdlWindow);
+#ifdef HELLOIMGUI_IS_DESKTOP_PLATFORM
         SDL_FlashWindow(sdlWindow, SDL_FLASH_UNTIL_FOCUSED);
         SDL_RaiseWindow(sdlWindow);
+#endif
     }
 
     ScreenBounds SdlWindowHelper::GetWindowBounds(WindowPointer window)
@@ -268,6 +278,13 @@ namespace HelloImGui { namespace BackendApi
         Uint32 flags = SDL_GetWindowFlags((SDL_Window  *)window);
         bool hidden = (flags & SDL_WINDOW_HIDDEN) != 0;
         return hidden;
+    }
+
+    void SdlWindowHelper::SetWindowTopMost(WindowPointer window, bool topMost)
+    {
+#ifdef HELLOIMGUI_IS_DESKTOP_PLATFORM
+        SDL_SetWindowAlwaysOnTop((SDL_Window *) window, topMost ? SDL_TRUE : SDL_FALSE);
+#endif
     }
 
     }} // namespace HelloImGui { namespace BackendApi