+++ src/hello_imgui/internal/docking_details.cpp
@@ -801,11 +801,12 @@ namespace DockingDetails
         ImGui::PushStyleVar(ImGuiStyleVar_WindowRounding, 0.0f);
         ImGui::PushStyleVar(ImGuiStyleVar_WindowBorderSize, 0.0f);
         ImGui::PushStyleVar(ImGuiStyleVar_WindowPadding, HelloImGui::EmToVec2(windowPaddingEm));
+        ImGui::PushStyleVar(ImGuiStyleVar_WindowMinSize, ImVec2(0.f, 0.f));
         if (windowBg.w != 0.f)
             ImGui::PushStyleColor(ImGuiCol_WindowBg, windowBg);
         static bool p_open = true;
-        ImGui::Begin(windowId.c_str(), &p_open, WindowFlagsNothing());
-        ImGui::PopStyleVar(3);
+        ImGui::Begin(windowId.c_str(), &p_open, WindowFlagsNothing() | ImGuiWindowFlags_NoScrollbar);
+        ImGui::PopStyleVar(4);
         if (windowBg.w != 0.f)
             ImGui::PopStyleColor();
         toolbarFunction();
@@ -845,11 +846,21 @@ namespace DockingDetails
 
         ImGui::PushStyleVar(ImGuiStyleVar_WindowRounding, 0.0f);
         ImGui::PushStyleVar(ImGuiStyleVar_WindowBorderSize, 0.0f);
-        ImGui::PushStyleVar(ImGuiStyleVar_WindowPadding, ImVec2(0.0f, 0.0f));
+        if (useDocking)
+        {
+            // When using docking, we want the window to have no padding, so that docked windows
+            // stick to the edges of the app window
+            ImGui::PushStyleVar(ImGuiStyleVar_WindowPadding, ImVec2(0.0f, 0.0f));
+        }
+
         static bool p_open = true;
         std::string windowTitle = useDocking ? "MainDockSpace" : "Main window (title bar invisible)";
         ImGui::Begin(windowTitle.c_str(), &p_open, WindowFlagsNothing());
-        ImGui::PopStyleVar(3);
+
+        if (useDocking)
+            ImGui::PopStyleVar(3);
+        else
+            ImGui::PopStyleVar(2);
     }
 
     void ImplProvideFullScreenImGuiWindow(const RunnerParams& runnerParams)